backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "insurance-outreach-backend"
version = "0.1.0"
description = "FastAPI backend for Insurance Outreach"
authors = [{ name = "Your Name" }]
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn>=0.29.0",
  "pydantic>=2.6.0",
  "python-dotenv>=1.0.0",
  "email-validator>=1.3.1",
  "requests>=2.26.0",
  "httpx>=0.24.0",
]

[tool.pytest.ini_options]
addopts = "-q --color=yes --cov=backend --cov-report=term-missing"
testpaths = ["tests"]
pythonpath = [".", "backend"]
